feat: 监控组循环检测开关 + interval 单参数模型
- 新建 CycleControlWidget：循环检测开关 + 条件显示 interval spinbox
- 5 面板（OCR/Image/Background/Number/Timed）统一替换
- 5 模块循环适配 cycle_enabled，关则单次执行后退出
- interval 改为 float（0=250-300ms 随机）
- pause 字段保留未用，向后兼容

+++ ui/main_window.py
@@ -406,7 +406,7 @@ class MainWindow(QMainWindow):
             if not panel or idx >= len(panel.groups_data):
                 QMessageBox.information(self, "测试", "组索引越界")
                 return
-            interval = int(panel.groups_data[idx].get(interval_key, default))
+            interval = float(panel.groups_data[idx].get(interval_key, default))
 
             wait_box = QMessageBox(self)
             wait_box.setWindowTitle("测试")
